requirements.txt
flask
numpy
pandas
xgboost
joblib
ipykernel
scikit-learn
seaborn
huggingface_hub
matplotlib
tensorflow
python-dotenv
google-generativeai